% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_charter.R
\name{bar_charter}
\alias{bar_charter}
\title{bar_charter}
\usage{
bar_charter(tab, var_title, doublebar = FALSE, dual_labels = NULL)
}
\arguments{
\item{tab}{table from twv_single or twv_mult function}

\item{var_title}{The title of the plot, an example would be "Blood Pressure"}

\item{doublebar}{This is an indication that there will be dual bars, and that the table
needs to be reformatted to a long version}

\item{dual_labels}{c("label1","label2") where these are the names for the legend

This makes bar charts for each risk value}
}
\description{
bar_charter
}
