% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subsetting.R
\name{demo_visit}
\alias{demo_visit}
\title{demo_visit}
\usage{
demo_visit(demo, status, spouse = TRUE)
}
\arguments{
\item{demo}{demographics data set}

\item{status}{visits data set}

\item{spouse}{True or False. Determines if the list returned contains spouse info
This function takes a demographics data set, a visit data set and creates a list
of the active employees and spouses. Spouses only returned it spouses parameter
is equal to TRUE}
}
\description{
demo_visit
}
