% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadDemo}
\alias{loadDemo}
\title{loadDemo}
\usage{
loadDemo(dir)
}
\arguments{
\item{dir}{This is the directory containg the data to load in

This is a fucntion that loads in demographics data set for a given company
The function takes the directory path way to the given company you would
like to load data for.}
}
\description{
loadDemo
}
