% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{loadStatus}
\alias{loadStatus}
\title{loadStatus}
\usage{
loadStatus(dir)
}
\arguments{
\item{dir}{path to directory
THis is a function that takes a directory path and returns the visit dataset
for a company in question.}
}
\description{
loadStatus
}
