% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_and_visit_tables.R
\name{outcome_cat_table}
\alias{outcome_cat_table}
\title{outcome_cat_table}
\usage{
outcome_cat_table(status, demo, date1, date2)
}
\arguments{
\item{status}{this is a data set of visits that has been through the sort_date
function so that it a subset between two dates}

\item{demo}{This is the demographics data set of all the active employees}

\item{date1}{lower bound date}

\item{date2}{upper bound date
This function takes a visit data set and a demo graphics data set and returns
a table of sums}
}
\description{
outcome_cat_table
}
