% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_subsetting.R
\name{sort_date}
\alias{sort_date}
\title{sort_date}
\usage{
sort_date(status, date1, date2)
}
\arguments{
\item{status}{visit data set. Should be the visit dataset from the out of demo_visit}

\item{date1}{lower boundary of date range}

\item{date2}{Upper boundary of date range

This is a function that takes a visit dataset (typically one that has been pushed through
the demo_visit function to only include active employees and spouses) returns a filtered
subset that is betweem the two date ranges. Both dates must be supplied.}
}
\description{
sort_date
}
