% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_with_values.R
\name{twv_combo}
\alias{twv_combo}
\title{twv_combo}
\usage{
twv_combo(current_vec, previous_vec)
}
\arguments{
\item{current_vec}{current visit vector}

\item{previous_vec}{previous visit vector

This function takes two  vectors of Risk values and creates a table of all
combinations of the risk values for each visit}
}
\description{
twv_combo
}
