% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_with_values.R
\name{twv_mult}
\alias{twv_mult}
\title{twv_mult}
\usage{
twv_mult(current_vec, previous_vec, new_vec = NULL)
}
\arguments{
\item{current_vec}{vector of risk value from most current visit}

\item{previous_vec}{vector of risk value from previous visit}

\item{new_vec}{vector of risk values for New Employees

This works in a similar manner as twv_singe, but it now makes tables with
current, previous, and new participant vectors}
}
\description{
twv_mult
}
