% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_with_values.R
\name{twv_single}
\alias{twv_single}
\title{twv_single
This is a function that makes a table with values (twv) for a
single vector of RIsk values. Typically this is the last visit
of a time period}
\usage{
twv_single(vec)
}
\arguments{
\item{vec}{is a vector of risk values}
}
\description{
twv_single
This is a function that makes a table with values (twv) for a
single vector of RIsk values. Typically this is the last visit
of a time period
}
