% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_and_visit_tables.R
\name{visit_table}
\alias{visit_table}
\title{visit_table}
\usage{
visit_table(status, date1, date2)
}
\arguments{
\item{status}{subsetted visit dates}

\item{date1}{lower date limit}

\item{date2}{upper date limit
This is a function that takes a subsetted visit date set from sort date
and returns the number of visits and number of first visits in a table}
}
\description{
visit_table
}
